+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+[1.0.10] - 2019-01-10
+
+Added
+  -
+
+Changed
+  -Column/commands.py
+  -Content/commands.py
+  -Foreignkey/commands.py
+  -MxNTable/commands.py
+  -Table/commands.py
+  -Makefile
+
+Deprecated
+  -
+
+Removed
+  -
+
+Fixed
+  -
+
+
 [1.0.9] - 2019-01-09
 
 Added